EspressoSimulations.jl Documentation Test

API Reference

Simulation Inputs

EspressoSimulations.EspressoParameters — Type

EspressoParameters(R0, L, a1, a2, 
                    ρ_out, ρ_grounds, 
                    csat, cs0, 
                    ϕs, Mout, P, 
                    Ds_star, Deff_star, k, 
                    tshot, vol_fracs)

Input parameters for the simulation that describe various parameters that influence the extraction.

Arguments

R0::Float64: Puck radius

L::Float64: Cylinder height

a1::Float64: fine grain radii [m]

a2::Float64: coarse grain radii [m]

ρ_out::Float64: density of water at 90 deg C [kg/m3]

ρ_grounds::Float64: density of roasted coffee bulk [kg/m3]

csat::Float64: saturation concentration of water outside grain. temperature dependent [kg/m3]

cs0::Float64: concentration of solubles in the grains initially [kg/m3], assuming preinfusion, see Heat and mass transfer in dispersed two-phase flows for pre-infusion details

ϕs::Float64: volume fraction of grounds in a packed bed [unitless] Heat and mass transfer in dispersed two-phase flows

Mout::Float64: mass of the beverage [kg]

P::Float64: pressure [bar]

Ds_star::Float64:

Deff_star::Float64:

k::Float64:

tshot::Float64:

vol_fracs::Float64:

Backend

EspressoSimulations.mass_matrix — Function

mass_matrix(N, Vi)

Generate a tile for a block mass matrix.
    
This differs by a factor of (1-ϕs) from the publication's supplementary [1]

mass_matrix(sim_params)

Build the mass matrix for our simulation from a set of 
`SimulationParameters`. This will generate a $2N^2+N$
block tridiagonal matrix. Block diagonal matrix using sub-blocks
from `mass_matrix(N, Vi)`.

This differs by a factor of (1-ϕs) from the publication's supplementary [1]

**Arguments**
- `sim_params::SimulationParameters`: simulation parameters to construct a mass matrix for

Post-processing

EspressoSimulations.extraction_percentage — Function

extraction_percentage(sol, dimensionless_params, sim_params; 
                        cumul=false)

Arguments:

sol::ODESolution

dimensionless_params::DimensionlessParameters

sim_params::SimulationParameters

cumul::Bool=false